feat: add note about slides --- title: Packaging in Python subtitle: The Good, the Bad, the Ugly! ---
This project contains my lecture notes for introductory and advance packaging with Python. There is a lot that _isn't_ covered — that's OK! There's always more to learn with packaging. The most important thing is to understand the basic components, and feel confident in trying things out!

:::{tip} Enhanced Slides Experience
Use [this gist] with the [Tampermonkey browser extension] to add {kbd}`a` and {kbd}`d` slide navigation.
:::

Everyone reading these notes are going to have a different starting point. Some of you will already be publishing packages. There may be a few of you who know how to author packages in an entirely different language altogether! Others will never have set foot in the world of Python packaging.

My aim is that everyone comes away from this workshop having learned _something_. I don't know ahead-of time what that might be for everyone!

:::{important} Goals of this workshop

- To end up with a Python package that can be published to PyPI.
- To know about the important "standards" and "tools" in packaging.
- To understand the difference between an _import_ package and _distribution_ package.
- To understand _how_ Python finds importable code.
- To have an idea of the _personas_ of your users.
  :::

[this gist]: https://gist.github.com/agoose77/570b303d66d061779db1e19e7f458976
[tampermonkey browser extension]: https://www.tampermonkey.net/